+++ packages/prebuild-config/src/plugins/unversioned/expo-splash-screen/getAndroidSplashConfig.ts
