+++ packages/prebuild-config/src/plugins/unversioned/expo-splash-screen/withAndroidSplashMainActivity.ts
@@ -1,8 +1,8 @@
 import { AndroidConfig, ConfigPlugin, withMainActivity } from '@expo/config-plugins';
-import { mergeContents } from '@expo/config-plugins/build/utils/generateCode';
+import { mergeContents, removeContents } from '@expo/config-plugins/build/utils/generateCode';
 import { ExpoConfig } from '@expo/config-types';
 
-import { getAndroidSplashConfig } from './getSplashConfig';
+import { getAndroidSplashConfig } from './getAndroidSplashConfig';
 
 export const withAndroidSplashMainActivity: ConfigPlugin = config => {
   return withMainActivity(config, config => {
@@ -21,12 +21,12 @@ export function setSplashScreenMainActivity(
   language: 'java' | 'kt'
 ): string {
   const splashConfig = getAndroidSplashConfig(config);
-  // TODO: Translucent is weird
-  const statusBarTranslucent = !!config.androidStatusBar?.translucent;
   if (!splashConfig) {
     // TODO: Remove splash screen code.
     return mainActivity;
   }
+  // TODO: Translucent is weird
+  const statusBarTranslucent = !!config.androidStatusBar?.translucent;
 
   const { resizeMode } = splashConfig;
   const isJava = language === 'java';
@@ -67,14 +67,34 @@ export function setSplashScreenMainActivity(
     }).contents;
   }
 
+  // DO NOT CHANGE
+  const showSplash = 'expo-splash-screen-mainActivity-onCreate-show-splash';
+
+  // Remove our generated code safely...
+  mainActivity = removeContents({
+    src: mainActivity,
+    tag: showSplash,
+  }).contents;
+
+  // Remove code from `@expo/configure-splash-screen`
+  mainActivity = mainActivity
+    .split('\n')
+    .filter(line => {
+      return !/SplashScreen\.show\(this,\s?SplashScreenImageResizeMode\./.test(line);
+    })
+    .join('\n');
+
+  // Reapply generated code.
   mainActivity = mergeContents({
     src: mainActivity,
     // insert just below super.onCreate
     anchor: /(?<=^.*super\.onCreate.*$)/m,
     offset: 1,
     comment: '//',
-    tag: 'expo-splash-screen-mainActivity-onCreate-show-splash',
-    newSrc: `          SplashScreen.show(this, SplashScreenImageResizeMode.${resizeMode.toUpperCase()}, ${statusBarTranslucent})${LE}`,
+    tag: showSplash,
+    newSrc: `    SplashScreen.show(this, SplashScreenImageResizeMode.${resizeMode.toUpperCase()}, ReactRootView${
+      isJava ? '.class' : '::class.java'
+    }, ${statusBarTranslucent})${LE}`,
   }).contents;
 
   // TODO: Remove old `SplashScreen.show`
